Implement Q9: Joint Particle Filter Observation
Includes implementation of observeUpdate() function for the
JointParticleFilter class of inference.py. This function weights and
resamples the entire list of particles based on the observation of all
ghost distances.

Had to rework the solution from Q8 to get this passing. The error from
the initial implementation of Q8 was the lack of randomization, and
possibly not considering a large enough distribution space.

Provisional grades
==================
Question q8: 1/1
------------------
Total: 1/1

Provisional grades
==================
Question q9: 3/3
------------------
Total: 3/3

Signed-off-by: MacKevin Fey <mackfey@cs.washington.edu>